+++ Macros/UnstableEnumMacro/+AbstractSourceLocation.swift
@@ -1,0 +1,8 @@
+import SwiftSyntaxMacros
+
+extension AbstractSourceLocation {
+    var description: String {
+        let file = self.file.description.filter({ ![" ", #"""#].contains($0) })
+        return "\(file):\(self.line.description)"
+    }
+}
