+++ Macros/UnstableEnumMacro/MacroError.swift
@@ -1,6 +1,6 @@
 import SwiftDiagnostics
 
-enum MacroError: Error {
+enum MacroError: String, Error {
     case isNotEnum
     case macroDoesNotHaveRequiredGenericArgument
     case unexpectedGenericArgument
@@ -14,40 +14,6 @@ enum MacroError: Error {
     case badEnumCaseTrailingTrivia
     case badEnumCaseComment
     case couldNotFindLocationOfNode
-    case cannotUnwrapSyntax(type: Any.Type, line: UInt = #line)
-
-    var rawValue: String {
-        switch self {
-        case .isNotEnum:
-            return "isNotEnum"
-        case .macroDoesNotHaveRequiredGenericArgument:
-            return "macroDoesNotHaveRequiredGenericArgument"
-        case .unexpectedGenericArgument:
-            return "unexpectedGenericArgument"
-        case .rawValuesNotAcceptable:
-            return "rawValuesNotAcceptable"
-        case .inconsistentQuotesAroundComment:
-            return "inconsistentQuotesAroundComment"
-        case .emptyValuesAreNotAcceptable:
-            return "emptyValuesAreNotAcceptable"
-        case .allEnumCasesWithIntTypeMustHaveACommentForValue:
-            return "allEnumCasesWithIntTypeMustHaveACommentForValue"
-        case .enumSeemsToHaveIntValuesButGenericArgumentSpecifiesString:
-            return "enumSeemsToHaveIntValuesButGenericArgumentSpecifiesString"
-        case .intEnumMustOnlyHaveIntValues:
-            return "intEnumMustOnlyHaveIntValues"
-        case .valuesMustBeUnique:
-            return "valuesMustBeUnique"
-        case .badEnumCaseTrailingTrivia:
-            return "badEnumCaseTrailingTrivia"
-        case .badEnumCaseComment:
-            return "badEnumCaseComment"
-        case .couldNotFindLocationOfNode:
-            return "couldNotFindLocationOfNode"
-        case let .cannotUnwrapSyntax(type, line):
-            return "cannotUnwrapSyntax(type: \(type), line: \(line))"
-        }
-    }
 }
 
 extension MacroError: DiagnosticMessage {
