+++ Macros/UnstableEnumMacro/RawKind.swift
@@ -1,0 +1,3 @@
+enum RawKind: String {
+    case String, Int, UInt
+}
